autonomous-research-agent/requirements.txt
langchain
openai
wikipedia
tavily-python
python-dotenv